+++ dhlab/utils/display.py
@@ -46,35 +46,7 @@ def printmd(S):
     display(Markdown(S))
     return
 
-def update(module="", overwrite=True, silent=False):
-    """Fetch modules from Github and write them to folder"""
-    nba = requests.get(
-        "https://raw.githubusercontent.com/Yoonsen/Modules/master/{module}.py".format(module=module),
-        headers={'Cache-Control': 'no-cache'}
-        )
-    filename = '{m}.py'.format(m=module)
-    if nba.status_code == 200:
-        file_exists = os.path.exists(filename)
-        if file_exists and not(overwrite):
-            if not silent:
-                printmd("File {f} exists - call `update('{m}', overwrite = True)` in order to download module `{m}` anyway".format(f = os.path.abspath(filename), m = module))
-        else:
-            nba = nba.text
-            with open(filename,'w', encoding='UTF-8') as pyfile:
-                pyfile.write(nba)
-                pyfile.flush()
-                pyfile.close()
-            if not silent:
-                printmd("Updated file `{module}.py`".format(module= os.path.abspath(module)))
-    else:
-        printmd( 
-            """{intro} for {module} with error {code}""".format(
-                intro = "An error occured during download", 
-                module = module, 
-                code= nba.status_code
-            )
-        )
-    return
+
 
 def css(url = "https://raw.githubusercontent.com/Yoonsen/Modules/master/css_style_sheets/nb_notebook.css"):
     """Associate a css stylesheet with the notebook, just specify a file or web reference, default is a custom css"""
